+++ Supplier_Task_with_Storage.h
@@ -4,9 +4,9 @@ class Supplier_Task_with_Storage :
 	public Supplier_Task
 {
 public:
-	Supplier_Task_with_Storage(int n, int m, int T,
-		const std::vector<int>& a, const std::vector<std::vector<int>>& b,
-		const std::vector<std::vector<int>>& C, const std::vector<std::set<int>>& D);
+	Supplier_Task_with_Storage(short int n, short int m, short int T,
+		const std::vector<short int>& a, const std::vector<std::vector<short int>>& b,
+		const std::vector<std::vector<short int>>& C, const std::vector<std::set<short int>>& D);
 	virtual int solve();
 	int getMinStorage() { return minStorage; }
 protected:
